src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-element-outside-shadow-style-expected.txt

CONSOLE ERROR: line 22: Uncaught TypeError: hostWithChild.createShadowRoot is not a function
